непоно

+++ Storage/yiiserver/controllers/StorageController.php
@@ -2,8 +2,10 @@
 
 namespace app\controllers;
 
+use app\models\DeleteId;
 use app\models\Files;
 use app\models\Storage;
+use Faker\Provider\File;
 
 class StorageController extends \yii\web\Controller
 {
@@ -13,6 +15,13 @@ class StorageController extends \yii\web\Controller
         if (isset($_GET['id'])) {
             if ($_GET['id'] == 0) {
                 $model = new Storage();
+                $delNode = DeleteId::find()->all();
+                if (count($delNode)) {
+                    $model->node_id = $delNode[0]->delete_id;
+                    $delNode[0]->delete();
+                } else {
+                    $model->node_id = count(Storage::find()->all()) + 1;
+                }
                 $model->setAttribute('time', time());
                 if ($model->save()) {
                     $model = Storage::find()->where(['time' => $model->time])->one();
@@ -53,8 +62,7 @@ class StorageController extends \yii\web\Controller
             $model->download = 0;
             $model->hash = md5_file("temp/" . $_POST['filename']);
             $model->count = count(Storage::find()->all());
-            if(!$model->save())
-            {
+            if (!$model->save()) {
                 echo "не сохранилось в базу";
                 return 0;
             }
@@ -62,8 +70,8 @@ class StorageController extends \yii\web\Controller
             $f = fopen("temp/" . $_POST['filename'], "rb");
             $length = filesize("temp/" . $_POST['filename']);
 
-            $dx = $length/count(Storage::find()->all());;
-            for ($i = 0; $i < ($length/$dx)-1; ) {
+            $dx = $length / count(Storage::find()->all());;
+            for ($i = 0; $i < ($length / $dx) - 1;) {
                 fseek($f, $i * $dx);
                 $bytes = fread($f, $dx);
                 $fw = fopen("temp/" . $_POST['filename'] . $i, 'wb');
@@ -72,7 +80,7 @@ class StorageController extends \yii\web\Controller
                 $i++;
             }
             fseek($f, $i * $dx);
-            $bytes = fread($f, $length-$i*$dx+1);
+            $bytes = fread($f, $length - $i * $dx + 1);
             $fw = fopen("temp\\" . $_POST['filename'] . $i, 'wb');
             fwrite($fw, $bytes);
             fclose($fw);
@@ -82,12 +90,30 @@ class StorageController extends \yii\web\Controller
         }
     }
 
+    public function actionDownload()
+    {
+        if(isset($_GET['filename']))
+        {
+            echo "http://yiiserver/temp/".$_GET['filename'].$_GET['id'];
+        }
+        else{
+            echo Files::find()->where(['id'=> count(Files::find()->all())] )->one()->name_file;
+        }
+
+    }
+
     public function deleteNotActiveNode()
     {
         $model = Storage::find()->all();
         foreach ($model as $node) {
             if (time() - $node->time >= 23) {
-                $node->delete();
+                $delId = new DeleteId();
+                $delId->delete_id = $node->node_id;
+                if ($delId->save()) {
+                    $node->delete();
+                } else {
+                    echo "не удалось удалить ноду";
+                }
             }
         }
     }
